Ripping out data generation, partially migrated to stylesheet usage
