--- data/generate-data.ts
@@ -1,3 +1,0 @@
-import { generateData } from "./src/generate-data";
-
-generateData();
