--- data/index.ts
@@ -1,1 +1,0 @@
-export { randomColor } from "./src/color";
