--- data/src/internal.ts
@@ -1,4 +1,0 @@
-export * from "./color";
-export * from "./constants";
-export * from "./data";
-export * from "./layerify";
