+++ src/config/cache.ts
@@ -1,0 +1,2 @@
+export const TILE_DATA_CDN_CACHE = 24 /* hrs */ * 60 * 60;
+export const TILE_DATA_CLIENT_CACHE = 3 /* hrs */ * 60 * 60;
