+++ src/hooks/mapbox-map.ts
@@ -1,18 +1,7 @@
 import { useState, useEffect, useCallback } from "react";
 import mapboxgl, { Map, MapEventType } from "mapbox-gl";
-import layerStyles from "~/data/generated/tax_parcels-layer-styles.json";
-import oldLayerStyles from "~/data/generated/tax_parcels_old-layer-styles.json";
-import publicLandStyles from "~/data/generated/public_land-layer-styles.json";
 import { useRouterWithHash } from "@/hooks/use-router-with-hash";
 import { mapboxToken } from "@/env";
-import {
-  HIGHLIGHTED_FILL_STYLE,
-  HIGHLIGHTED_LINE_STYLE,
-  SELECTED_FILL_STYLE,
-  SELECTED_LINE_STYLE,
-  getLineWidth,
-} from "@/config/styles";
-import { getHighlightLayer, getSelectedLayer } from "@/lib/layers";
 
 export const LAND_LAYERS = ["tax_parcels", "public_land", "tax_parcels_old"];
 
@@ -21,8 +10,6 @@ export const useMapboxMap = () => {
   const [mapInitialized, setMapInitialized] = useState(false);
   const [mapContainer, mapContainerRef] = useState<HTMLDivElement | null>(null);
   const router = useRouterWithHash();
-  const [selectionLayers, setSelectionlayers] = useState<string[]>([]);
-  const [highlightLayers, setHighlightLayers] = useState<string[]>([]);
   const [layers, setLayers] = useState<string[]>([]);
 
   const onMapRender = useCallback((e: MapEventType["render"]) => {
@@ -65,74 +52,14 @@ export const useMapboxMap = () => {
         // To trigger events that are waiting on tile data
         _map.on("render", onMapRender);
 
-        _map.addSource("LCMDParcels", {
-          type: "vector",
-          tiles: [
-            `${window.location.protocol}${window.location.host}/tiles/{z}/{x}/{y}`,
-          ],
-          maxzoom: 16,
-        });
-
         // Layers
-        const { highlightedLayers, selectedLayers } = addSelectableLayers(
-          [
-            {
-              id: "tax_parcels",
-              slot: "bottom",
-              source: "LCMDParcels",
-              "source-layer": "tax_parcels",
-              type: "fill",
-              // @ts-expect-error
-              paint: layerStyles,
-            },
-            {
-              id: "tax_parcels_old",
-              slot: "bottom",
-              source: "LCMDParcels",
-              "source-layer": "tax_parcels_old",
-              type: "fill",
-              // @ts-expect-error
-              paint: oldLayerStyles,
-              layout: {
-                visibility: "none",
-              },
-            },
-            {
-              id: "public_land",
-              source: "LCMDParcels",
-              slot: "bottom",
-              "source-layer": "public_land",
-              type: "fill",
-              // @ts-expect-error
-              paint: publicLandStyles,
-            },
-            {
-              id: "Eastside_Reroutes",
-              source: "LCMDParcels",
-              "source-layer": "eastside_reroutes",
-              type: "line",
-              paint: {
-                "line-color": [
-                  "case",
-                  ["boolean", ["feature-state", "selected"], false],
-                  "#48c242",
-                  "#594630",
-                ],
-                "line-width": getLineWidth({ unselectedSize: -30 }),
-              },
-            },
-          ],
-          _map
-        );
-        setSelectionlayers(selectedLayers);
-        setHighlightLayers(highlightedLayers);
         setLayers([
           "Eastside_Reroutes",
           "tax_parcels",
           "public_land",
           "tax_parcels_old",
         ]);
-        addHillShade(_map);
+        // May need to setTerrrain ?
 
         // Events
 
@@ -146,90 +73,6 @@ export const useMapboxMap = () => {
     mapContainer,
     mapContainerRef,
     mapInitialized,
-    selectionLayers,
-    highlightLayers,
     layers,
-  };
-};
-
-const addHillShade = (map: Map) => {
-  map.addSource("mapbox-dem-hillshade", {
-    type: "raster-dem",
-    url: "mapbox://mapbox.mapbox-terrain-dem-v1",
-    maxzoom: 13,
-  });
-  map.addSource("mapbox-dem", {
-    type: "raster-dem",
-    url: "mapbox://mapbox.mapbox-terrain-dem-v1",
-    maxzoom: 13,
-  });
-  map.addLayer({
-    id: "hillshading",
-    source: "mapbox-dem-hillshade",
-    type: "hillshade",
-    paint: {
-      "hillshade-exaggeration": 0.2,
-    },
-  });
-  map.setTerrain({ source: "mapbox-dem", exaggeration: 2.5 });
-};
-
-const addSelectableLayers = (layers: mapboxgl.AnyLayer[], map: Map) => {
-  const layerResults = layers.map((layer) =>
-    addSelectableLayer({
-      layer,
-      map,
-      order: layer.type === "line" ? "below" : "above",
-    })
-  );
-  return layerResults.reduce(
-    (result, { highlightedLayer, selectedLayer }) => {
-      return {
-        highlightedLayers: [...result.highlightedLayers, highlightedLayer],
-        selectedLayers: [...result.selectedLayers, selectedLayer],
-      };
-    },
-    { highlightedLayers: [], selectedLayers: [] } as {
-      highlightedLayers: string[];
-      selectedLayers: string[];
-    }
-  );
-};
-
-const addSelectableLayer = ({
-  layer,
-  map,
-  order = "above",
-}: {
-  layer: mapboxgl.AnyLayer;
-  map: Map;
-  order?: "above" | "below";
-}) => {
-  if (order === "above") {
-    map.addLayer(layer);
-  }
-  if (layer.type === "fill" || layer.type === "line") {
-    const { paint, layout, ...unstyledLayer } = layer;
-    map.addLayer({
-      ...unstyledLayer,
-      type: "line",
-      id: getSelectedLayer(layer.id),
-      ...(layer.type === "fill" ? SELECTED_FILL_STYLE : SELECTED_LINE_STYLE),
-    });
-    map.addLayer({
-      ...unstyledLayer,
-      type: "line",
-      id: getHighlightLayer(layer.id),
-      ...(layer.type === "fill"
-        ? HIGHLIGHTED_FILL_STYLE
-        : HIGHLIGHTED_LINE_STYLE),
-    });
-  }
-  if (order === "below") {
-    map.addLayer(layer);
-  }
-  return {
-    highlightedLayer: `${layer.id}_highlighted`,
-    selectedLayer: `${layer.id}_selected`,
   };
 };
