+++ src/types/properties.ts
